requirements.txt
pyyaml==6.0.2
supabase==2.11.0
dash==2.18.2
pandas==2.2.3